requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
astroid==2.5.6
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
orjson==3.5.3
pep8==1.7.1
promise==2.3
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pylint==2.8.3
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.17
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wrapt==1.12.1
